// packages/vue3components/src/Table/src/types/table.ts
export declare type SortOrder = 'ascend' | 'descend'

export interface TableCurrentDataSource<T = Recordable> {
  currentDataSource: T[]
}

export interface TableRowSelection<T = any> extends ITableRowSelection {
  /**
   * Callback executed when selected rows change
   * @param selectedRowKeys 已选的 keyValues
   * @param selectedRows 已选的 records
   * @param isClickCustomRow 是否是点击行触发（反之，就是点击checkbox/radiobox）
   * @returns void
   */
  onChange?: (
    selectedRowKeys: Key[],
    selectedRows: T[],
    isClickCustomRow?: boolean
  ) => void

  /**
   * Callback executed when select/deselect one row
   * @type Function
   */
  onSelect?: (
    record: T,
    selected: boolean,
    selectedRows: Object[],
    nativeEvent: Event
  ) => any

  /**
   * Callback executed when select/deselect all rows
   * @type Function
   */
  onSelectAll?: (selected: boolean, selectedRows: T[], changeRows: T[]) => any

  /**
   * Callback executed when row selection is inverted
   * @type Function
   */
  onSelectInvert?: (selectedRows: Key[]) => any
}

export interface ExpandedRowRenderRecord<T> extends TableCustomRecord<T> {
  indent?: number
  expanded?: boolean
}
export interface ColumnFilterItem {
  text?: string
  value?: string
  children?: any
}

export interface TableCustomRecord<T = Recordable> {
  record?: T
  index?: number
}

export interface SorterResult {
  column: ColumnProps
  order: SortOrder
  field: string
  columnKey: string
}

export interface FetchParams {
  searchInfo?: Recordable
  page?: number
  sortInfo?: Recordable
  filterInfo?: Recordable
}

export interface GetColumnsParams {
  ignoreIndex?: boolean
  ignoreAction?: boolean
  sort?: boolean
}

export type SizeType = 'default' | 'middle' | 'small' | 'large'
// ...
export interface FetchSetting {
  // 请求接口当前页数
  pageField: string
  // 每页显示多少条
  sizeField: string
  // 请求结果列表字段  支持 a.b.c
  listField: string
  // 请求结果总数字段  支持 a.b.c
  totalField: string
}

export interface TableSetting {
  redo?: boolean
  size?: boolean
  setting?: boolean
  settingCache?: boolean
  fullScreen?: boolean
}

export interface BasicTableProps<T = any> {
// ...
  rowKey?: InstanceType<typeof AntDesignVueTable>['$props']['rowKey']
// ...
  scroll?: InstanceType<typeof AntDesignVueTable>['$props']['scroll']
// ...
}

export type CellFormat =
  | string
  | ((text: string, record: Recordable, index: number) => string | number)
  | Map<string | number, any>

// @ts-ignore
export interface BasicColumn extends ColumnProps<Recordable> {
  children?: BasicColumn[]
  filters?: {
    text: string
    value: string
    children?:
      | unknown[]
      | (((props: Record<string, unknown>) => unknown[]) &
          (() => unknown[]) &
          (() => unknown[]))
  }[]

  //
  flag?: 'INDEX' | 'DEFAULT' | 'CHECKBOX' | 'RADIO' | 'ACTION'
  customTitle?: VueNode

  slots?: Recordable

  // 自定义header渲染
  customHeaderRender?: (
    column: BasicColumn
  ) => string | VNodeChild | JSX.Element
  // Whether to hide the column by default, it can be displayed in the column configuration
  defaultHidden?: boolean

  // Help text for table column header
  helpMessage?: string | string[] | VNodeChild | JSX.Element

  format?: CellFormat

  // Editable
  edit?: boolean
  editRow?: boolean
  editable?: boolean
  editComponent?: ComponentType
  editComponentProps?:
    | ((opt: {
        text: string | number | boolean | Recordable
        record: Recordable
        column: BasicColumn
        index: number
      }) => Recordable)
    | Recordable
  editRule?: boolean | ((text: string, record: Recordable) => Promise<string>)
  editValueMap?: (value: any) => string
  onEditRow?: () => void
  // 权限编码控制是否显示
  auth?: string | string[]
  // 业务控制是否显示
  ifShow?: boolean | ((column: BasicColumn) => boolean)
  // 自定义修改后显示的内容
  editRender?: (opt: {
    text: string | number | boolean | Recordable
    record: Recordable
    column: BasicColumn
    index: number
    currentValue: string | number | boolean | Recordable
  }) => VNodeChild | JSX.Element
  // 动态 Disabled
  editDynamicDisabled?: boolean | ((record: Recordable) => boolean)
}